Move types to extra_body module
