+++ tensorzero-internal/src/inference/types/extra_body.rs
@@ -1,6 +1,17 @@
 use serde::{Deserialize, Serialize};
+use serde_json::Value;
 
-use super::InferenceExtraBody;
+#[derive(Clone, Debug, Default, Deserialize, PartialEq, Serialize)]
+#[serde(transparent)]
+pub struct ExtraBodyConfig {
+    pub data: Vec<ExtraBodyReplacement>,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+pub struct ExtraBodyReplacement {
+    pub pointer: String,
+    pub value: Value,
+}
 
 /// The 'InferenceExtraBody' options provided directly in an inference request
 /// These have not yet been filtered by variant name
@@ -34,4 +45,41 @@ impl UnfilteredInferenceExtraBody {
 #[serde(transparent)]
 pub struct FilteredInferenceExtraBody {
     pub data: Vec<InferenceExtraBody>,
+}
+
+/// Holds the config-level and inference-level extra body options
+#[derive(Clone, Debug, Default, PartialEq, Serialize)]
+pub struct FullExtraBodyConfig {
+    pub extra_body: ExtraBodyConfig,
+    pub inference_extra_body: FilteredInferenceExtraBody,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+#[serde(untagged)]
+pub enum InferenceExtraBody {
+    Provider {
+        model_provider_name: String,
+        pointer: String,
+        value: serde_json::Value,
+    },
+    Variant {
+        variant_name: String,
+        pointer: String,
+        value: serde_json::Value,
+    },
+}
+
+impl InferenceExtraBody {
+    pub fn should_apply_variant(&self, variant_name: Option<&str>) -> bool {
+        match (self, variant_name) {
+            (InferenceExtraBody::Provider { .. }, _) => true,
+            (
+                InferenceExtraBody::Variant {
+                    variant_name: v, ..
+                },
+                Some(expected_name),
+            ) => v == expected_name,
+            (InferenceExtraBody::Variant { .. }, None) => false,
+        }
+    }
 }
